refactor: move makePromise to core
