+++ functions/package.json
@@ -20,8 +20,6 @@
     "@mauriciorobayo/gods-promises": "^2.0.2",
     "@typescript-eslint/experimental-utils": "^4.28.0",
     "axios": "^0.21.1",
-    "bible-passage-reference-parser": "^2.0.1",
-    "bible-reference-formatter": "^1.0.1",
     "firebase-admin": "^9.2.0",
     "firebase-functions": "^3.11.0",
     "jsdom": "^16.6.0",
